// src/app/core/http/generated/cart/v1/model/cartLineItem.ts
/**
 * kart-cart-service API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LineItemAvailability } from './lineItemAvailability';


export interface CartLineItem { 
    sku: string;
    quantity: number;
    availability: LineItemAvailability;
}
export namespace CartLineItem {
}